backend/requirements.txt
fastapi[standard]
uvicorn
python-dotenv
requests
python-multipart
pypdf
python-docx
pdf2image
pytesseract
pillow
